refactor(trusty-mpm): split spawn_disclaim mod.rs to clear 500-SLOC cap
mod.rs grew to 558 SLOC after adding disclaimed_stderr_piped_spawn and
disclaimed_stdout_piped_spawn in the previous commit, tripping CI's
"500-line file-size cap" job (a repo-specific lint job not covered by
fmt/clippy/test, so it wasn't run before that commit landed).

Split by cohesion, following the pattern the macos::stderr_piped /
macos::stdout_piped modules already established: each primitive's
public API (struct, handle enum, spawn fn, tests) now lives in its own
top-level sibling module (stderr_piped.rs, stdout_piped.rs) instead of
inline in mod.rs. The macOS posix_spawn implementations already had
their own files; this pulls the platform-agnostic half out to match.

Pure code-organization change, no behavior delta: the public API
(disclaimed_stderr_piped_spawn, disclaimed_stdout_piped_spawn,
StderrPipedSpawn, StdoutPipedSpawn) is re-exported from mod.rs at the
same crate::core::spawn_disclaim:: path via `pub use`, so no caller
changes. mod.rs is now 259 SLOC.

Verified: cargo fmt/clippy/test all still pass at both crate and full
workspace scope, and scripts/check_line_cap.sh reports 0 violations.
